implemented logout
