+++ app/components/navbar/TopNav.jsx
@@ -1,93 +1,64 @@
 import React, { Component } from 'react';
-import RegForm from '../user/Reg';
-import LoginForm from '../user/Login';
-import Modal from 'react-modal';
+import LoginRegButton from '../user/LoginRegButton';
+import LogoutButton from '../user/LogoutButton';
 
 export default class TopNav extends Component {
   constructor(props) {
     super(props);
+
     this.state = {
-      regModalIsOpen: false,
-      loginModalIsOpen: false
+      isLoggedIn: false
     }
   }
   
-  // reset state
-  resetState = () => {
-    this.setState({
-      regModalIsOpen: false,
-      loginModalIsOpen: false
+  // check for sessions on page refresh
+  componentDidMount(){
+    $.ajax({
+      url: 'http://localhost:8080/checkifloggedin',
+      dataType: 'json',
+      type: 'POST',
+      xhrFields: { withCredentials: true },
+      success: (result) => {
+        this.setState({isLoggedIn: result.isLoggedIn});
+      },
+      error: (err) => {
+        console.error(err);
+      }
     });
   }
 
-  regOpenModal = () => {
-    this.setState({ regModalIsOpen: true });
+  // reset state
+  resetState = () => {
+    this.setState({
+      isLoggedIn: false
+    });
   }
 
-  regCloseModal = () => {
-    this.resetState();
+  handleLoginSession = () => {
+    this.setState({ isLoggedIn: true });
   }
 
-  loginOpenModal = () => {
-    this.setState({ loginModalIsOpen: true });
-  }
-
-  loginCloseModal = () => {
-    this.resetState();
+  handleLogoutSession = () => {
+    this.setState({ isLoggedIn: false });
   }
 
   render() {
-    const modalStyles = {
-      content: {
-        width: '700px',
-        padding: '30px',
-        top: '50%',
-        left: '50%',
-        right: 'auto',
-        bottom: 'auto',
-        marginRight: '-50%',
-        transform: 'translate(-50%, -50%)'
-      }
-    };
+    const isLoggedIn = this.state.isLoggedIn;
 
     return (
-      <nav className="navbar navbar-toggleable-md navbar-inverse fixed-top bg-inverse">
+      <nav className="topnav navbar navbar-toggleable-md navbar-inverse fixed-top bg-inverse">
         <button className="navbar-toggler navbar-toggler-right hidden-lg-up" type="button" data-toggle="collapse" data-target="#navbarsExampleDefault" aria-controls="navbarsExampleDefault" aria-expanded="false" aria-label="Toggle navigation">
         <span className="navbar-toggler-icon"></span>
         </button>
         <a className="navbar-brand" href="/">Dashboard</a>
 
         <div className="collapse navbar-collapse" id="navbarsExampleDefault">
-          <ul className="navbar-nav ml-auto">
-            <li className="nav-item">
-              <button className="btn btn-secondary" onClick={ this.regOpenModal }>Registration</button>
-              <Modal
-                isOpen={ this.state.regModalIsOpen }
-                onRequestClose={ this.regCloseModal }
-                style={ modalStyles }
-                contentLabel="Reg Modal"
-              >
-                <h3 className="pl-0 d-flex modal-header">Registration: <i className="fa fa-times justify-content-right" onClick={ this.regCloseModal } /></h3>
-                <RegForm close={ this.regCloseModal } />
-              </Modal>
-            </li>
 
-            <li className="nav-item">
-              <button className="btn btn-secondary" onClick={ this.loginOpenModal }>Login</button>
-              <Modal
-                isOpen={ this.state.loginModalIsOpen }
-                onRequestClose={ this.loginCloseModal }
-                style={ modalStyles }
-                contentLabel="Reg Modal"
-              >
-                <h3 className="pl-0 d-flex modal-header">Login: <i className="fa fa-times justify-content-right" onClick={ this.loginCloseModal } /></h3>
-                <LoginForm close={ this.loginCloseModal } />
-              </Modal>
-            </li>
+            { isLoggedIn ? (<LogoutButton handleLogoutSession={this.handleLogoutSession}/>) : (<LoginRegButton handleLoginSession={this.handleLoginSession} />) }
 
-          </ul>
         </div>
       </nav>
     );
   }
 }
+
