+++ app/components/user/RegForm.jsx
@@ -1,10 +1,11 @@
 import React, { Component } from 'react';
 
-export default class LoginForm extends Component {
+export default class RegForm extends Component {
   constructor(props) {
     super(props);
     this.state = {
       username: '',
+      email: '',
       password: ''
     }
   }
@@ -20,25 +21,28 @@ export default class LoginForm extends Component {
 
     let formData = {
       username: this.state.username,
+      email: this.state.email,
       password: this.state.password,
     }
 
     // error checking
-    if (formData.username.length < 1 || formData.password.length < 1) {
+    if (formData.username.length < 1 || formData.email.length < 1 || formData.password.length < 1) {
       return false;
     }
 
     $.ajax({
-      url: 'http://localhost:8080/login',
+      url: 'http://localhost:8080/register',
       dataType: 'json',
       type: 'POST',
       data: formData,
-      success: (result) => {
+      xhrFields: { withCredentials: true },
+      success: (data) => {
         this.props.close();
-        alert('Logged in');
+        this.props.handleLoginSession();
+        alert('Registration successful');
       },
-      error: (xhr, status, err) => {
-        console.error(status, err);
+      error: (err) => {
+        console.error('Error', err);
         alert('There was some problem with the form. Please resubmit');
       }
     });
@@ -46,6 +50,7 @@ export default class LoginForm extends Component {
     // reset state after form submission
     this.setState({
       username: '',
+      email: '',
       password: '',
     });
   };
@@ -58,16 +63,23 @@ export default class LoginForm extends Component {
             User Name:
           </label>
           <input id="formUser" className="form-control" name='username' type="text"
-            onChange={ this.handleKeyChange('username') } />
+            onChange={this.handleKeyChange('username')} />
+        </div>
+        <div className="form-group row pl-3 pr-3">
+          <label className="col-form-label-sm">
+            Email:
+          </label>
+          <input id="formEmail" className="form-control" name='email' type="email" id="email"
+            onChange={this.handleKeyChange('email')} />
         </div>
         <div className="form-group row pl-3 pr-3">
           <label className="col-form-label-sm">
             Password:
           </label>
           <input id="formPassword" className="form-control" name='password' type="password" id="password"
-            onChange={ this.handleKeyChange('password') } />
+            onChange={this.handleKeyChange('password')} />
         </div>
-        <button className="btn btn-primary" type="submit">Login</button>
+        <button className="btn btn-primary pull-right" type="submit">Sign up</button>
       </form>
     );
   }
