+++ styles/application.scss
@@ -17,4 +17,4 @@ $card-shadow-color: grey;
 @import 'modal';
 @import 'dashboard';
 @import 'tooltip';
-
+@import 'topnav';
